Update cqlsh and cqlshlib / test folder
Update core `cqlsh` files to `cassandra` 3.11.10:
* https://github.com/apache/cassandra/tree/cassandra-3.11.10/pylib
* https://github.com/apache/cassandra/blob/cassandra-3.11.10/bin/cqlsh.py

The following files are updated:
* `cqlsh` using the contents of `cqlsh.py`. It's confusing that this
file doesn't match the upstream `cqlsh`, but upstream is a shell script
entrypoint for picking the python, which we don't need since we're using
`setuptools` to manage that. This is consistent with how things have
been done in the past.
* `cqlshlib` folder content is updated via a direct copy of upstream
  `cassandra` repo, except for the `test` folder which is pulled up a
  directory so that it won't be installed in a normal install.
* `test` folder is a direct copy of what's in upstream `cassandra`,
except for moving the location up one folder.