Drop test stuff, it makes no sense in a re-packaging
There's no need to have the `test` folder as this is a repackaging of
upstream. Anyone wanting to run unit tests/develop against the real
`cqlsh` should use the test infra upstream in `cassandra`, as they
include docker containers, entrypoint scripts, etc. So there's no need
for us to deal with the overhead of maintaining it here.